+++ src/api/controllers/pvp/index.spec.js
@@ -85,11 +85,6 @@ describe('pvp controller', () => {
 
   describe('leaderboard', () => {
     const apiTokenId = 3;
-    const user = testData.user({
-      tokenId: apiTokenId,
-      alias: 'madou',
-      accountName: 'madou.1234',
-    });
 
     const createStanding = (rank) => testData.dbStanding({
       apiTokenId,
@@ -117,53 +112,15 @@ describe('pvp controller', () => {
       four,
     ];
 
-    const assertIsSorted = (arr, key) => {
-      const fromOneArr = arr.slice(1);
-      fromOneArr.forEach((item, index) => {
-        const prevItem = arr[index];
-        expect(item[key] >= prevItem[key]).to.be.true;
-      });
-    };
-
-    const assertUsers = (leaderboard) => {
-      leaderboard.forEach((standing) => {
-        expect(standing).to.include({
-          alias: user.alias,
-          accountName: user.accountName,
-        });
-      });
-    };
-
-    beforeEach(() => {
-      readUser.withArgs(models, { apiTokenId }).returns(Promise.resolve(user));
-    });
-
     ['na', 'gw2a', 'eu'].forEach((region) => {
       beforeEach(() => {
         listPvpStandings.withArgs(models, season.id, region).returns(Promise.resolve(standings));
       });
 
       describe(region, () => {
-        it('should return maxium of 250', async () => {
-          const longStandings = Array(300).fill(one);
-          listPvpStandings
-            .withArgs(models, season.id, region)
-            .returns(Promise.resolve(longStandings));
-          const leaderboard = await controller.leaderboard(region);
-
-          expect(leaderboard.length).to.equal(250);
-        });
-
-        it('should build leaderboard and sort by highest to lowest rating', async () => {
-          const leaderboard = await controller.leaderboard(region);
-
-          assertIsSorted(leaderboard, `${region}Rank`);
-        });
-
         it('should add user data to each standing', async () => {
           const leaderboard = await controller.leaderboard(region);
-
-          assertUsers(leaderboard);
+          expect(leaderboard).to.equal(standings);
         });
       });
     });