+++ src/fetch/fetchers/pvpLeaderboard.js
@@ -6,7 +6,7 @@ import _ from 'lodash';
 
 import gw2, { readLatestPvpSeason } from 'lib/gw2';
 import { bulkCreateStubUser } from 'lib/services/user';
-import { saveList as saveStandings, list as listStandings } from 'lib/services/pvpStandings';
+import { saveList as saveStandings, list as listPvpStandings } from 'lib/services/pvpStandings';
 import createLogger from 'lib/gitter';
 import buildLadderByAccountName from '../lib/leaderboard';
 
@@ -77,7 +77,7 @@ export default async function calculatePvpLeaderboards (models: Models) {
   const [naLadder, euLadder, standings] = await Promise.all([
     gw2.readPvpLadder(null, season.id, { region: 'na' }),
     gw2.readPvpLadder(null, season.id, { region: 'eu' }),
-    listStandings(models, season.id),
+    listPvpStandings(models, season.id),
   ]);
 
   const newUsersResults = await addMissingUsers(models, naLadder.concat(euLadder));